calculador: agrega componente de carga de producto y refactoriza products/[slug]
