+++ src/app/(app)/dashboard/products/components/LoadingPage.tsx
@@ -1,0 +1,9 @@
+import MyProgress from '@/components/MyProgress'
+
+export default function LoadingPage() {
+  return (
+    <div className="flex h-screen w-screen items-center justify-center">
+      <MyProgress />
+    </div>
+  )
+}
